# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiErrorMessage {
  error: string;
}

export interface AnalysisJobCreated {
  analysis_id: string;
}

export type AnalysisJobStatusStatus = typeof AnalysisJobStatusStatus[keyof typeof AnalysisJobStatusStatus];


export const AnalysisJobStatusStatus = {
  pending: 'pending',
  done: 'done',
  error: 'error',
} as const;

export interface NoesisRegion {
  id: string;
  label: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  x: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  y: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  width: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  height: number;
  sequence_order: number;
  explanation: string;
  why_it_matters: string;
  related_region_ids: string[];
  relationship_explanation: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
}

export interface NoesisAnalysis {
  title: string;
  image_type: string;
  central_question: string;
  overall_summary: string;
  big_takeaway: string;
  regions: NoesisRegion[];
}

export interface AnalysisJobStatus {
  status: AnalysisJobStatusStatus;
  analysis?: NoesisAnalysis;
  error?: string;
}

export interface AnalysisInput {
  /**
     * Data URL (base64) of the uploaded image, e.g. data:image/png;base64,...
     * @minLength 30
     */
  image_data_url: string;
}