# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analysis
    description: Visual analysis operations
paths:
  /analyze:
    post:
      operationId: analyzeImage
      tags: [analysis]
      summary: Start analysis of an uploaded visual
      description: Starts AI vision analysis of an uploaded image and returns a job id to poll (analysis can exceed proxy request timeouts, so results are fetched via polling)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisInput"
      responses:
        "202":
          description: Analysis started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisJobCreated"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
        "502":
          description: Analysis could not be started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
  /analyze/{analysisId}:
    get:
      operationId: getAnalysisStatus
      tags: [analysis]
      summary: Poll analysis job status
      description: Returns the status of an analysis job, including the walkthrough result once done
      parameters:
        - name: analysisId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisJobStatus"
        "404":
          description: Unknown or expired job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ApiErrorMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AnalysisJobCreated:
      type: object
      properties:
        analysis_id:
          type: string
      required:
        - analysis_id
    AnalysisJobStatus:
      type: object
      properties:
        status:
          type: string
          enum: [pending, done, error]
        analysis:
          $ref: "#/components/schemas/NoesisAnalysis"
        error:
          type: string
      required:
        - status
    AnalysisInput:
      type: object
      properties:
        image_data_url:
          type: string
          description: Data URL (base64) of the uploaded image, e.g. data:image/png;base64,...
          minLength: 30
      required:
        - image_data_url
    NoesisRegion:
      type: object
      properties:
        id: { type: string }
        label: { type: string }
        x: { type: number, minimum: 0, maximum: 1 }
        y: { type: number, minimum: 0, maximum: 1 }
        width: { type: number, minimum: 0, maximum: 1 }
        height: { type: number, minimum: 0, maximum: 1 }
        sequence_order: { type: number }
        explanation: { type: string }
        why_it_matters: { type: string }
        related_region_ids:
          type: array
          items: { type: string }
        relationship_explanation: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
      required:
        - id
        - label
        - x
        - y
        - width
        - height
        - sequence_order
        - explanation
        - why_it_matters
        - related_region_ids
        - relationship_explanation
        - confidence
    NoesisAnalysis:
      type: object
      properties:
        title: { type: string }
        image_type: { type: string }
        central_question: { type: string }
        overall_summary: { type: string }
        big_takeaway: { type: string }
        regions:
          type: array
          items:
            $ref: "#/components/schemas/NoesisRegion"
      required:
        - title
        - image_type
        - central_question
        - overall_summary
        - big_takeaway
        - regions