# lib/api-zod/src/generated/types/analysisJobStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisJobStatusStatus } from './analysisJobStatusStatus';
import type { NoesisAnalysis } from './noesisAnalysis';

export interface AnalysisJobStatus {
  status: AnalysisJobStatusStatus;
  analysis?: NoesisAnalysis;
  error?: string;
}